Remove marshal method and class names from CoreCLR output
